requirements.txt
django
requests
djangorestframework
django-cors-headers
django-filter
djangorestframework-simplejwt
drf-yasg # documentation